requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pandas>=2.2.0
numpy>=1.26.0
psutil>=5.9.0